add utilities file, dress up app
