+++ __init__.py
@@ -1,0 +1,2 @@
+import scripts
+from scripts import utils
